requirements.txt
requests>=2.31.0
keyboard>=0.13.5
mss>=9.0.1
opencv-python>=4.9.0
psutil>=5.9.0
numpy>=1.26.0
pytesseract>=0.3.10
Pillow>=10.0.0
